move back thehive_test to test/alerters and add the unit test for the hive description in thehive_test.py
